requirements.txt
# VigilAI Core Dependencies

# AI/ML Framework
tensorflow==2.15.0
torch==2.1.0
torchvision==0.16.0
transformers==4.35.0
opencv-python==4.8.1.78
mediapipe==0.10.7
scikit-learn==1.3.2
numpy==1.24.3
scipy==1.11.4

# Data Processing
pandas==2.1.3
apache-kafka==2.0.2
pyspark==3.5.0
delta-spark==3.0.0

# Edge Computing
jetson-inference==1.0.0
jetson-utils==1.0.0

# Hardware Integration
pyserial==3.5
pybluez==0.23
obd==0.7.0

# Web Framework
fastapi==0.104.1
uvicorn==0.24.0
websockets==12.0

# Monitoring & Logging
prometheus-client==0.19.0
grafana-api==1.0.3
elasticsearch==8.11.0

# Security
cryptography==41.0.7
pycryptodome==3.19.0

# Testing
pytest==7.4.3
pytest-asyncio==0.21.1
pytest-cov==4.1.0
locust==2.17.0

# Cloud Infrastructure
kubernetes==28.1.0
helm==3.12.0
docker==6.1.3
terraform==1.6.0

# Streaming & Messaging
kafka-python==2.0.2
aiokafka==0.8.11
pydantic-kafka==0.1.0

# Monitoring & Observability
prometheus-client==0.19.0
grafana-api==1.0.3
jaeger-client==4.8.0
opentelemetry-api==1.20.0
opentelemetry-sdk==1.20.0
opentelemetry-instrumentation-fastapi==0.41b0

# Analytics & BI
plotly==5.17.0
dash==2.14.1
superset==3.0.0
jupyter==1.0.0
mlflow==2.7.1

# Security
vault==0.1.0
cert-manager==1.13.0
falco==0.1.0
opa==0.1.0

# CI/CD
github-actions==0.1.0
jenkins==0.1.0
argo-cd==0.1.0

# Mobile App Dependencies
react-native==0.72.0
expo==49.0.0
react-navigation==6.1.0
react-native-chart-kit==6.12.0
react-native-vector-icons==10.0.0
react-native-gesture-handler==2.12.0
react-native-reanimated==3.4.0

# Accessibility Features
pyttsx3==2.90
speechrecognition==3.10.0
gtts==2.4.0
pygame==2.5.0
opencv-python==4.8.1.78
pillow==10.0.1

# Emergency Response
twilio==8.5.0
smtplib==0.1.0
email==0.1.0

# System Monitoring
psutil==5.9.6
scipy==1.11.4
numpy==1.24.3

# Database & Storage
psycopg2-binary==2.9.7
sqlalchemy==2.0.21
alembic==1.12.0

# Additional AI/ML
scikit-learn==1.3.2
joblib==1.3.2
tensorflow==2.15.0
torch==2.1.0
transformers==4.35.0

# Development
black==23.11.0
flake8==6.1.0
mypy==1.7.1
pre-commit==3.5.0
coverage==7.3.2
xmlrunner==3.2.0
